workout_api/requirements.txt
# Django and REST Framework
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.0.0
django-filter>=23.2

# Database
psycopg2-binary>=2.9.6  # PostgreSQL adapter (for production)

# Computer Vision and AI
opencv-python>=4.8.0
# mediapipe>=0.10.0  # Temporarily commented - will install manually
numpy>=1.24.0
Pillow>=10.0.0

# Video processing
# moviepy>=1.0.3  # Temporarily commented - has dependency issues

# Environment variables
python-decouple>=3.8

# Development tools
pytest>=7.4.0
pytest-django>=4.5.2
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0

# Production
gunicorn>=21.2.0
whitenoise>=6.5.0

# Utilities
python-dateutil>=2.8.2
requests>=2.31.0